refactor(auth): use JWT token from backend for authentication
- Store JWT access token in localStorage instead of user data
- Decode JWT token to extract userId, userName, and isAdmin from payload
- Update LoginPage to receive and store token from backend response
- Update App.jsx to decode token on page reload for session persistence
- Update axiosConfigure to clear authToken on refresh token failure
- Fix import paths in AdminRoute and PrivateRoute
- Update tests to mock jwt-decode and expect token in response

Security improvement: User data now comes from signed JWT payload,
making it tamper-proof. Modified tokens will fail signature validation
on the backend.
